refactor(node-resolution): modularize node module resolution logic
Break down the node module resolution function into smaller, focused helper functions to improve readability and maintainability. This includes creating a context object, separating package.json handling, and isolating export path resolution logic
