+++ tests/node-resolution/node-module-resolution-1.test.tsx
@@ -1,0 +1,32 @@
+import { describe, expect, test } from "bun:test"
+import { runTscircuitCode } from "lib/runner"
+
+describe("node module resolution", () => {
+  test("resolves package.json", async () => {
+    const circuitJson = await runTscircuitCode(
+      {
+        "node_modules/test-package/package.json": JSON.stringify({
+          name: "test-package",
+        }),
+        "node_modules/test-package/index.js": `
+          export const resistorName = "R1"
+          export const resistanceValue = "1k"
+        `,
+        "user-code.tsx": `
+          import { resistorName, resistanceValue } from "test-package"
+          export default () => (<resistor name={resistorName} resistance={resistanceValue} />)
+        `,
+      },
+      {
+        mainComponentPath: "user-code",
+      },
+    )
+
+    const resistor = circuitJson.find(
+      (element) => element.type === "source_component" && element.name === "R1",
+    ) as any
+    expect(resistor).toBeDefined()
+    expect(resistor.resistance).toBe(1000)
+    expect(resistor.name).toBe("R1")
+  })
+})
